% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{bu_save}
\alias{bu_save}
\title{Save Dataframe to .rds or .csv with Backup File}
\usage{
bu_save(x, full_path, compress = c("none", "gz"))
}
\arguments{
\item{x}{A dataframe}

\item{full_path}{The full path containing the folder and filename with file extension}

\item{compress}{Only available for .rds files (one of 'none' or 'gz')}
}
\description{
Saves a dataframe to the specified location and creates another backup file with 
date-time extension (yy-mm-dd_HH-MM-SS)
}
\examples{
library(RFtex)

df <- tibble::tibble(col = c(1,2,3))
bu_save(df, "test_df.rds", "none")
}
