% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{dir_create}
\alias{dir_create}
\title{Create directories}
\usage{
dir_create(dirs, rec = TRUE)
}
\arguments{
\item{dirs}{A character string of full path dirnames}

\item{rec}{Shall dirs be created recursively? (default: TRUE)}
}
\description{
Create directories
}
\examples{
dirs <- c("test/testdir1", "test/testdir2", "test/testdir1")
dir_create(dirs)
list.dirs("test/", recursive = FALSE)
system(paste0("rm -r ", "test/"))

}
