% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_generation.R
\name{id_chunk}
\alias{id_chunk}
\title{GENERATE IDs for chunk splits}
\usage{
id_chunk(num, len, sample = FALSE)
}
\arguments{
\item{num}{Number of chunks}

\item{len}{Length of the chunk vector}

\item{sample}{Shall the IDs be randomized}
}
\value{
An integer vector
}
\description{
GENERATE IDs for chunk splits
}
\examples{
id_chunk(5, 28, FALSE)
id_chunk(5, 28, TRUE)
}
