% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_generation.R
\name{id_split}
\alias{id_split}
\title{GENERATE IDs for splits}
\usage{
id_split(num, len, sample = FALSE)
}
\arguments{
\item{num}{Number of splits}

\item{len}{Length of the split vector}

\item{sample}{Shall the IDs be randomized}
}
\value{
An integer vector
}
\description{
GENERATE IDs for splits
}
\examples{
id_split(5, 28, FALSE)
id_split(5, 28, TRUE)
}
