% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{pbs}
\alias{pbs}
\title{Profile Based Standardization (PBS)}
\usage{
pbs(table, clus = 1)
}
\arguments{
\item{table}{A dataframe with 4 columns (columns need to be in the right order)\cr
column 1: A document identifier\cr
column 2: The concept id\cr
column 3: A term identifier\cr
column 4: The term frequency}

\item{cl}{Number of cores to be used}
}
\value{
A dataframe with showing the PBS for each unique firm-pair for each concept
}
\description{
Profile Based Standardization (PBS)
}
\examples{
library (RFtex)

table <- tibble::tibble(
doc_id = c(1,1,1,1,1,2,2,2,2,3,3,3,3),
con_id = c(1,1,2,2,2,1,1,3,3,1,3,2,3),
term   = c("a", "b", "c", "d", "e", "a", "b", "f", "g", "a", "f", "e", "g"),
freq   = c(21,12,58,32,14,21,14,66,14,12,85,100,12))

pbs(table)
}
